feat: directed acyclic graph + context closure set reduction tooling
context closure:
- composable context rules that close over a context object and allow
  for a resolve / fail system.
- designed to remove items from sets that fail the context closure for
  the set, to replace part of the current Feature / Service dependency
  system in Dairy Core.

DAG:
- composable DAG rules
- makes it easy to add dependencies to any node in the system
- includes a kahn sort to detect cycles while sorting
- replaces the other part of the current Feature / Service dependency
  system in Dairy Core

Both systems are capable of working with `Set`s of any type, using a
selector function to pull out the rule on each.

This allows for a more flexible system, with some types having several
different adjacency and context rules, which can then be selected by the
sorter of the graph.